requirements.txt
Django==5.2.1
djangorestframework==3.15.1
django-cors-headers==4.3.1
